D4P # 347 Public Portal - Unable to submit a Claim and Duplicate File Uploads Allowed When Submitting Standard Claim via OpenShift Public Portal (#904)
* Added @angular/animations and compiler-cli missing packages to fix the build errors

* Fixed all the components to use standalone: false after 19 upgrade

* Updated angular material and angular CDK libraries

* Updated all the kendo-angular libraries to version 19

* Updated ngx-mask library to version 19

* - Fixed the issues with supportCategory, supportSubCategory in the location service
- Updated the ng-recaptcha-2 version to 15
- Updated docker file to support angular 19

* D4P# 143 Portal - Public Claim Appeals - Launching Appeal Process Modal
Fixed the styles and button on the confirm appeal dialog model

* D4P# 211 Portal - Public Claim Appeals - New Claim Decision Screen
- Added Mat stepper component to the claim decision
- Added logic to remove click functionality on the stepper
- Added styles to stepper component

* D4p# 212 Portal - Public Claim Appeals - Appeal Button Visibility and Functionality
Added dfa_bpfclosedate, claimtype, isadjustmentclaim, late appealallowed feilds to recovery claim entity
Updated the GetClaimDetailsAsync to get these new feilds
Added mappings to these new fields
Added logic to enable the appeals button
Added logic to disable and show the hover when the 60 days is crossed.
Added logic to verify the lateAppealAllowed is true to enable the button.
Added logic to calculate the 60 days from date of decision

* D4P # 27 PORTAL - Private Appeals - New timeline style and content

Started investigating the existing get application query to get the list of appeals

Added appeal model to the to get the case Id, Appeal status, Appeal stage and Appeal Id

Updated the mapping with the new fields

* D$P # 347 Public Portal - Unable to submit a Claim and Duplicate File Uploads Allowed When Submitting Standard Claim via OpenShift Public Portal
- Fixed the issue with date dateFileClosed null validation issue