D4P #  27 PORTAL - Private Appeals - New timeline style and content (#919)
* Added @angular/animations and compiler-cli missing packages to fix the build errors

* Fixed all the components to use standalone: false after 19 upgrade

* Updated angular material and angular CDK libraries

* Updated all the kendo-angular libraries to version 19

* Updated ngx-mask library to version 19

* - Fixed the issues with supportCategory, supportSubCategory in the location service
- Updated the ng-recaptcha-2 version to 15
- Updated docker file to support angular 19

* D4P# 143 Portal - Public Claim Appeals - Launching Appeal Process Modal
Fixed the styles and button on the confirm appeal dialog model

* D4P# 211 Portal - Public Claim Appeals - New Claim Decision Screen
- Added Mat stepper component to the claim decision
- Added logic to remove click functionality on the stepper
- Added styles to stepper component

* D4p# 212 Portal - Public Claim Appeals - Appeal Button Visibility and Functionality
Added dfa_bpfclosedate, claimtype, isadjustmentclaim, late appealallowed feilds to recovery claim entity
Updated the GetClaimDetailsAsync to get these new feilds
Added mappings to these new fields
Added logic to enable the appeals button
Added logic to disable and show the hover when the 60 days is crossed.
Added logic to verify the lateAppealAllowed is true to enable the button.
Added logic to calculate the 60 days from date of decision

* D4P # 27 PORTAL - Private Appeals - New timeline style and content

Started investigating the existing get application query to get the list of appeals

Added appeal model to the to get the case Id, Appeal status, Appeal stage and Appeal Id

Updated the mapping with the new fields

* D$P # 347 Public Portal - Unable to submit a Claim and Duplicate File Uploads Allowed When Submitting Standard Claim via OpenShift Public Portal
- Fixed the issue with date dateFileClosed null validation issue

* Fixed the issue with the appeals button logic used for testing

* Adjusted the appeal controller  to send appeal by Id
Updated the routes on appeals to see wether its create or view
Updated the viewAppealAfterSubmission method to include the new route
Updated Appeal service to get te appeal model
Populate the Appeal Reason on the  view appeals
Added appeal ID to the routing

* Added shared- Ui to the package-json file

* D4P# 27 PORTAL - Private Appeals - New timeline style and content
- Added DFA_CaseEligibilityAppeal and DFA_CasePaidAmountAppeal entities using the XRM tool
- Updated the entities, Option sets and database context files

* D4P# 27 PORTAL - Private Appeals - New timeline style and content
Added dateappealdecisionmade on the entities, gateway, ,mappers, Application controller
Generated Frontend Models for the new fields

* D4P #  27 PORTAL - Private Appeals - New timeline style and content
Added IAppealRepository and Mapper to dynamics gateway
Updated the GetApplicationListAsync method to load the list of apppeals and the case eligibility appeal
- Updated CaseEligibilityAppeal and CasePaidAmountAppeal to the appeal entities, currentcaseAppeal Model, Appeal
- Updated the cuurent case appeal model
- Added Appeal, dfa_appeal mappings
- generated frontend API services
- Added DT models to appeals, caseEligibility Appeal, CAsePAidAmount Appeal,
- Added database mappings to DFA_CaseEligibilityAppeal, ProcessStage, CaseAppealComposite
- Added Appeal query to the Appeal repositoty
_ Added Unit test for the Appeal query

* D4P #  27 PORTAL - Private Appeals - New timeline style and content
Added logic to get the amount appeal to dynamics
Added CasePaidAmountAppeal to the currentcaseappeal,  mappings, models
- Added new mapping for the DFA_CasePaidAmountAppeal and CasePaidAmountAppeal
- Updated  the dfa-application componnet to show the ststus of the appeal
- Added new QueryAmountPaidAppeal to get the amount appeals loaded.

* Fixed the comments in dynamics gateway

* Updated the appeal mapper to parse string to Guid