D4P# 25 : PORTAL/RAFT - Private Appeals - New Appeal Screens and flow (#975)
* D4P# 25 PORTAL - Private Appeals - New Appeal Screens and flow
Added Applicant Type to GetApplicationMainById
- Added new api method to get case details by caseID
- Added the GetCaseDetailsListAsync method to dynamics gateway, handler and Idynamicsgateway
- Added dfa Application Type field
- Updated appeal create endpoint
- Added new mappings for application type and also case details
Generated the frontend API models
- Updated dfa appeal component with the loader functionality
- Updateed the functionality on how we get application data, case details and also how to load the mat stepper steps on appeal
- Added new routes to send the appeal Id
- Added functionality to create the appeal o the model popup button
- Updated casedetails component to get data from the application main and case details
- Added applicatio
ndetails and acse details to the appeal stepper componnets
Added new routes for view and edit the appeals on the router component

* D4P# 25 PORTAL - Private Appeals - New Appeal Screens and flow
Added update endpoint for appeals
Added updatedappealrequest model
Added AppealUpdateRequest mappings
Generated front end models for the Update functionality
Added back to dashboard button on appeals page
Added snackbar to show that the appeal has successfully created.
Added logic to view the eligibility and amount appeals
Updated supporting documents to use appeal id not the case id

* Fixed the issue with the uploading documents by passing the correct appeal id
Added snackbar to show when the appeal is successfully submitted
Added the logic to show the success message on the view only mode on sign and submit
Added logic to disable the upload documents button and delete icon on view only
Added logic to diaable the reason field on view only

* Added GetSignatureAnnotation method to retreive the signature from dynamics
- Added GetAnnotationByFileNameAsync to dataverse layer
- Updated the signature model to use the exsisting signature block
- Updated the GetAppeal method to load the signature
-Added signeture block model mappings
- Removed the signature block model (using exsisting one)
- Added the initial signature, datesigned, signedname to the sign and submit appeals form
- Updated sign and submit ts file to get the signature fields
- Updated drawsignature to append with base64 if the input doesnot have that
- Updated the getAppealSignAndSubmitForm in the dfaappeal component

* Generated the latest dynamics entities

* D4P # 357 PORTAL - Private Appeals - Viewing Submitted Appeals
- Added new feilds for sign, signature and date in appeal controller
- Refacter all the functions to use this new feilds instead of the sign and submit models
- Refactored the signature upload and get functions
- Updated the mappings with these three new fields
- Updated the XRM tool box to get the new fields from dynamics
- Created the FE models for the new changes to the appeal models
- deleted the update DTO model and directly using the model in the submit appeal functionality
- Updated the logic to wireup  new feilds on the appeal component
- Added successfully submitted message on the sign and submit page
- Updated the appeal DTO with new fields and updated teh appeal mappings

* D4P # 387 PORTAL - Private Appeals: Display App Application's Portal Status & Portal Note based on the Private App Application BPF Stage & Checklist Item
- Removed the errorpage navigation on all the componnets
- Added snackbar to show the error message when there is an issue instead of redirecting to an error page
- Add null check for the portalnotes in dynamics gateway

* D4P # PORTAL - Private Appeals - Viewing Submitted Appeals
Hide navigating the buttons on the the view appeal workflow
Fixed the issues with the appeal buttons and timeline not showing on the dfa application page
Changed the viewAppealAfterSubmission method to use the appeal ID as the appeals array is deleted
Fixed the styles on the BPF stages

* D4P # 387 PORTAL - Private Appeals: Display App Application's Portal Status & Portal Note based on the Private App Application BPF Stage & Checklist Item
Added Createdinversion and the portal status note to the DFA application
- Added mappings to these new fields
- Added these new fields to GetApplicationListAsync in dynamics gateway
- Added CaseBPFVersionSet Enum type
- Generated fronend models to these new fields

* D4P # 387 PORTAL - Private Appeals: Display App Application's Portal Status & Portal Note based on the Private App Application BPF Stage & Checklist Item
-Added new BPF for the application status
-Updated teh code to show the portal status
- Fixed the old application status bar to show the createdonportal status
- Added newApplicationItems items arrya to show the new BPF stages
- Added backend logic to show the BPF stages

* D4P # 387 PORTAL - Private Appeals: Display App Application's Portal Status & Portal Note based on the Private App Application BPF Stage & Checklist Item

Fixed the merge conflits in dynamics gateway and mapper
Fixed all the issues with showing the appeals button and appeal timelines
added condition to load the eligibility appeal and amount appeal

* D4P # 387 PORTAL - Private Appeals: Display App Application's Portal Status & Portal Note based on the Private App Application BPF Stage & Checklist Item
- Fixed the issue with the stepper component not navigating correctly
- Fixed the styles with the new timeline and appeals timeline
- Fixed the View button styles to view the appeals

* D4P # 387 PORTAL - Private Appeals: Display App Application's Portal Status & Portal Note based on the Private App Application BPF Stage & Checklist Item
- Added logic to create multiple appeals
When there  is eligibility appeal overturned then able to create amount appeal
Fixed the buttons for the view appeals

* D4P # 25 PORTAL - Private Appeals - New Appeal Screens and flow
Fixed the issue with the cancel appeal dialog issue.

* D4P # 25 PORTAL - Private Appeals - New Appeal Screens and flow
Fixed the issue with the cancel appeal dialog issue.

* D4P # 25 PORTAL - Private Appeals - New Appeal Screens and flow and D4P # 357
Fixed the issue with showing the Add support documents on edit and view workflows
Added functionality on the appeal reason page to show the eligible and appeal sections.
Fixed the issue with showing the appeal type correctly on the dfa appeal component
Fixed the issue with not showing the cancel button on appeal creation and also hide the return to dashboard page

* Fixed the typo

* D4P # 28 PORTAL - Private Appeals - Appeal document uploading and tagging
- Added the confirmation snackbar after the document is submitted.

* D4P # 24 PORTAL - Private Appeals - Case Confirmation
Fixed the issue with the createdon date not showing on the appeal
created mapping for createdon
Added enum conversion for the appplicant type

* Added application createdon to current application
Added mappings for the creationdate
Generated frontend models
Updared the original application date to the new createddate

D4P# 25 : PORTAL/RAFT - Private Appeals - New Appeal Screens and flow
Updated the text on the amount appeal component
removed the radiobuttons and added checkbox for the confirmreport

* d4P # 25 PORTAL/RAFT - Private Appeals - New Appeal Screens and flow
Generated the dynamics entities using the XRMtool box
Added the new ReviewedEvaluatorReport to appeals model, mappings, controller create and update model,
Fixed the frontend to save the ReviewedEvaluatorReport form field